implemented review authorization

+++ routes/campgrounds.js
@@ -1,21 +1,11 @@
 const express = require('express');
 const router = express.Router();
 const catchAsync = require('../utils/catchAsync');
-const ExpressError = require('../utils/ExpressError');
 const Campground = require('../models/campground');
-const {campgroundSchema} = require('../schemas.js');
-const  {isLoggedIn} = require('../middleware');
+const  {isLoggedIn, isAuthor, validateCampground} = require('../middleware');
 
 
-const validateCampground = (req,res,next) => {
-    const {error} = campgroundSchema.validate(req.body)
-    if(error){
-        const message = error.details.map(el => el.message).join(',')
-        throw new ExpressError(message, 400)
-    } else {
-        next();
-    }
-}
+
 
 // router.get('/', (req,res) => {
 
@@ -41,7 +31,12 @@ router.post('/', isLoggedIn, validateCampground, catchAsync(async (req,res,next)
 
 router.get('/:id', catchAsync(async (req,res,next) => {
     const {id} = req.params;
-    const campground = await Campground.findById(id).populate('reviews').populate('author');
+    const campground = await Campground.findById(id).populate({
+        path: 'reviews',
+        populate: {
+            path: 'author'
+        }
+    }).populate('author')
     if(!campground){
         req.flash('error', 'Cannot find that campground.')
         return res.redirect('/campgrounds')
@@ -50,7 +45,7 @@ router.get('/:id', catchAsync(async (req,res,next) => {
 }))
 
 
-router.get('/:id/edit', isLoggedIn, catchAsync(async(req,res,next) => {
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async(req,res,next) => {
     const {id} = req.params;
     const campground = await Campground.findById(id);
     if(!campground){
@@ -60,15 +55,20 @@ router.get('/:id/edit', isLoggedIn, catchAsync(async(req,res,next) => {
     res.render('campgrounds/edit', {campground})
 }))
 
-router.put('/:id', isLoggedIn, validateCampground, catchAsync(async(req,res,next) => {
+router.put('/:id',  isLoggedIn, isAuthor, validateCampground, catchAsync(async(req,res,next) => {
     const {id} = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, {...req.body.campground});
+    const campground = await Campground.findByIdAndUpdate(id,{...req.body.campground})
     req.flash('success', 'Successfully updated the campground!');
     res.redirect(`/campgrounds/${campground._id}`)
 }))
 
-router.delete('/:id', catchAsync(async(req,res,next) => {
+router.delete('/:id',  isLoggedIn, isAuthor, catchAsync(async(req,res,next) => {
     const {id} = req.params;
+    const campground = await Campground.findById(id);
+    if(!campground.author.equals(req.user._id)){
+        req.flash('error', 'You do not have permission to do that!');
+        return res.redirect(`/campgrounds/${campground._id}`);
+    }
     await Campground.findByIdAndDelete(id)
     req.flash('success', 'Successfully deleted the campground.')
